+++ internal/app/handlers/errors.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"errors"
-	"github.com/stlesnik/url_shortener/cmd/logger"
+	"github.com/stlesnik/url_shortener/internal/logger"
 	"net/http"
 )
