+++ internal/app/middleware/logging.go
@@ -1,7 +1,7 @@
 package middleware
 
 import (
-	"github.com/stlesnik/url_shortener/cmd/logger"
+	"github.com/stlesnik/url_shortener/internal/logger"
 	"net/http"
 	"time"
 )
